==> 1_upload_data.ipynb <==
This notebook helps to upload data from local to hugging face!

Imports

from datasets import Dataset, load_dataset, Audio
import pandas as pd
import os

Load Audio Data function

def load_audio_data(audio_dir):
    audio_files = []
    file_names = []
    
    # Iterate through audio directory
    for filename in os.listdir(audio_dir):
        if filename.endswith('.wav'):# Adjust this based on your audio file format
            # Load audio file
            audio_path = os.path.join(audio_dir, filename)
            audio_files.append(audio_path)

            file_names.append(filename)
    return {
        'audio_path': audio_files,
        'file_names': file_names,
    }

Upload Chinese

Upload chinese_b1 and chinese_b2 together. DONT do below chinese_b1 and chinese_b2 if doing this

b1_audio_dir = "data/recordings/chinese/chinese_b1"
b1_audio_data = load_audio_data(b1_audio_dir)
b1_audio_df = pd.DataFrame(b1_audio_data)

b1_df = pd.read_csv("data/csv/chinese_b1.csv")
b1_df['file_names'] = b1_df['file_names'].apply(lambda x: x + '.wav')
b1_df = pd.merge(b1_df, b1_audio_df, on='file_names', how='inner')

b2_audio_dir = "data/recordings/chinese/chinese_b2"
b2_audio_data = load_audio_data(b2_audio_dir)
b2_audio_df = pd.DataFrame(b2_audio_data)

b2_df = pd.read_csv("data/csv/chinese_b2.csv")
b2_df = pd.merge(b2_df, b2_audio_df, on='file_names', how='inner')

#merge b1 and b2
df = pd.concat([b1_df, b2_df], ignore_index=True)
df = df.drop(columns=['remarks'])
df = df.dropna(subset=['fluency','spoken_text'])

dict_data = df.to_dict(orient='list')
audio_dataset = Dataset.from_dict(dict_data).cast_column("audio_path", Audio())

audio_dataset.train_test_split(test_size=0.2).push_to_hub('chinese_children_speech')

chinese_b1

b1_audio_dir = "data/recordings/chinese/chinese_b1"
b1_audio_data = load_audio_data(b1_audio_dir)
b1_audio_df = pd.DataFrame(b1_audio_data)

b1_df = pd.read_csv("data/csv/chinese_b1.csv")
b1_df['file_names'] = b1_df['file_names'].apply(lambda x: x + '.wav')
b1_df = pd.merge(b1_df, b1_audio_df, on='file_names', how='inner')

b1_df = b1_df.drop(columns=['remarks'])
b1_df = b1_df.dropna(subset=['fluency','spoken_text'])
dict_data = b1_df.to_dict(orient='list')

audio_dataset = Dataset.from_dict(dict_data).cast_column("audio_path", Audio())
audio_dataset.train_test_split(test_size=0.2).push_to_hub('chinese_b1')

chinese_b2

audio_dir = "data/recordings/chinese/chinese_b2"
audio_data = load_audio_data(audio_dir)
df = pd.DataFrame(audio_data)

b2_df = pd.read_csv("data/csv/chinese_b2.csv")
df = pd.merge(df, b2_df, on='file_names', how='inner')

dict_data = df.to_dict(orient='list')
audio_dataset = Dataset.from_dict(dict_data).cast_column("audio_path", Audio())

audio_dataset.train_test_split(test_size=0.2).push_to_hub('chinese_b2')

Upload Malay

Malay batch 1

audio_dir = "data/recordings/malay/malay_batch1"
audio_data = load_audio_data(audio_dir)
df = pd.DataFrame(audio_data)

b1_df = pd.read_csv("data/csv/malay_batch1.csv")
b1_df = b1_df.drop(columns=['remarks'])
df = pd.merge(df, b1_df, on='file_names', how='inner')

dict_data = df.to_dict(orient='list')
audio_dataset = Dataset.from_dict(dict_data).cast_column("audio_path", Audio())

audio_dataset.train_test_split(test_size=0.2).push_to_hub('malay_batch1')

Upload Tamil

Upload ALL tamil together.

b1_audio_dir = "data/recordings/tamil/tamil_b1"
b1_audio_data = load_audio_data(b1_audio_dir)
b1_audio_df = pd.DataFrame(b1_audio_data)

b1_df = pd.read_csv("data/csv/tamil_b1.csv")
b1_df = pd.merge(b1_df, b1_audio_df, on='file_names', how='inner')

b2_audio_dir = "data/recordings/tamil/tamil_b2"
b2_audio_data = load_audio_data(b2_audio_dir)
b2_audio_df = pd.DataFrame(b2_audio_data)

b2_df = pd.read_csv("data/csv/tamil_b2.csv")
b2_df = pd.merge(b2_df, b2_audio_df, on='file_names', how='inner')

b3_audio_dir = "data/recordings/tamil/tamil_b3"
b3_audio_data = load_audio_data(b3_audio_dir)
b3_audio_df = pd.DataFrame(b3_audio_data)

b3_df = pd.read_csv("data/csv/tamil_b3.csv")
b3_df = pd.merge(b3_df, b3_audio_df, on='file_names', how='inner')

b4_all_audio_dir = "data/recordings/tamil/tamil_b4_all"
b4_all_audio_data = load_audio_data(b4_all_audio_dir)
b4_all_audio_df = pd.DataFrame(b4_all_audio_data)

b4_all_df = pd.read_csv("data/csv/tamil_b4_all.csv")
b4_all_df = pd.merge(b4_all_df, b4_all_audio_df, on='file_names', how='inner')

b4_p1_audio_dir = "data/recordings/tamil/tamil_b4_p1"
b4_p1_audio_data = load_audio_data(b4_p1_audio_dir)
b4_p1_audio_df = pd.DataFrame(b4_p1_audio_data)

b4_p1_df = pd.read_csv("data/csv/tamil_b4_p1.csv")
b4_p1_df = pd.merge(b4_p1_df, b4_p1_audio_df, on='file_names', how='inner')

b4_p2_audio_dir = "data/recordings/tamil/tamil_b4_p2"
b4_p2_audio_data = load_audio_data(b4_p2_audio_dir)
b4_p2_audio_df = pd.DataFrame(b4_p2_audio_data)

b4_p2_df = pd.read_csv("data/csv/tamil_b4_p2.csv")
b4_p2_df = pd.merge(b4_p2_df, b4_p2_audio_df, on='file_names', how='inner')

b5_audio_dir = "data/recordings/tamil/tamil_b5"
b5_audio_data = load_audio_data(b5_audio_dir)
b5_audio_df = pd.DataFrame(b5_audio_data)

b5_df = pd.read_csv("data/csv/tamil_b5.csv")
b5_df = pd.merge(b5_df, b5_audio_df, on='file_names', how='inner')

#merge b1 and b2
df = pd.concat([b1_df, b2_df,b3_df, b4_all_df,b4_p1_df, b4_p2_df,b5_df], ignore_index=True)
df = df.drop(columns=['remarks'])
df = df.dropna(subset=['fluency','spoken_text'])

dict_data = df.to_dict(orient='list')
audio_dataset = Dataset.from_dict(dict_data).cast_column("audio_path", Audio())

audio_dataset.train_test_split(test_size=0.2).push_to_hub('tamil_children_speech')
